# src/main/resources/swagger.yml
openapi: 3.0.3
info:
  title: PlayBall API
  description: |
    **PlayBall** - 스포츠 매칭 플랫폼 REST API 명세서

    ## 인증 방식
    카카오 OAuth 로그인 후 서버에서 JWT를 발급합니다.
    - `POST /api/auth/kakao` → 카카오 인가코드로 로그인 → accessToken + refreshToken 발급
    - 인증 필요 API: `Authorization: Bearer {accessToken}` 헤더 포함
    - 비로그인 허용 API: `security: []` 표기

    ## 공통 응답
    ```json
    { "success": true, "message": "성공", "data": { ... } }
    ```

    ## 에러 응답
    ```json
    { "success": false, "message": "에러 메시지", "data": null }
    ```

    ## 종목 코드
    `SOCCER` | `BASKETBALL` | `RUNNING` | `BADMINTON`

    ## 실력 코드
    `BEGINNER` | `INTERMEDIATE` | `ADVANCED`
  version: 1.1.0
  contact:
    name: PlayBall Team
servers:
  - url: http://localhost:8080
    description: 로컬 개발
  - url: https://api.playball.kr
    description: 운영
...
security:
  - BearerAuth: []

# ============================================================
paths:
...
  /api/auth/kakao:
    post:
      tags: [인증]
      summary: 카카오 로그인
      description: |
        카카오 인가코드를 전달하면 JWT 토큰을 발급합니다.
        - 기존 회원: `isNewUser: false` → 바로 사용
        - 신규 회원: `isNewUser: true` → `POST /api/members/signup/complete` 호출 필요
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [authorizationCode]
              properties:
                authorizationCode:
                  type: string
            example:
              authorizationCode: "ABCDEF123456"
      responses:
        '200':
          description: 로그인 성공 (신규/기존 회원 모두 200. isNewUser로 구분)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              examples:
                existingUser:
                  summary: 기존 회원
                  value:
                    success: true
                    message: "로그인 성공"
                    data:
                      accessToken: "eyJhbGciOiJIUzI1NiJ9..."
                      refreshToken: "dGhpcyBpcyByZWZyZXNo..."
                      tokenType: "Bearer"
                      isNewUser: false
                      member:
                        memberId: 1
                        nickname: "홍길동"
                        profileImage: "https://k.kakaocdn.net/..."
                        role: "USER"
                newUser:
                  summary: 신규 회원 (추가정보 입력 필요)
                  value:
                    success: true
                    message: "추가 정보 입력이 필요합니다"
                    data:
                      accessToken: "eyJhbGciOiJIUzI1NiJ9..."
                      refreshToken: "dGhpcyBpcyByZWZyZXNo..."
                      tokenType: "Bearer"
                      isNewUser: true
                      member:
                        memberId: 2
                        nickname: null
                        profileImage: "https://k.kakaocdn.net/..."
                        role: "USER"
        '401':
          description: 카카오 인증 실패
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                success: false
                message: "카카오 인증에 실패했습니다."
                data: null
  /api/auth/refresh:
    post:
      tags: [인증]
      summary: 토큰 재발급
      description: Refresh Token으로 새 Access Token + Refresh Token을 발급받습니다.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
            example:
              refreshToken: "dGhpcyBpcyByZWZyZXNo..."
      responses:
        '200':
          description: 재발급 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "토큰 재발급 성공"
                data:
                  accessToken: "eyJhbGciOiJIUzI1NiJ9.new..."
                  refreshToken: "bmV3IHJlZnJlc2g..."
                  tokenType: "Bearer"
                  isNewUser: false
        '401':
          description: 만료되었거나 유효하지 않은 Refresh Token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                invalid:
                  value: { success: false, message: "유효하지 않은 토큰입니다.", data: null }
                expired:
                  value: { success: false, message: "만료된 토큰입니다.", data: null }
  /api/auth/logout:
    post:
      tags: [인증]
      summary: 로그아웃
      description: 서버에 저장된 Refresh Token을 삭제합니다.
      responses:
        '200':
          description: 로그아웃 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "로그아웃 성공"
                data: null
        '401':
          $ref: '#/components/responses/Unauthorized'

  # ──────────────────────────────────────────────
  # 회원
  # ──────────────────────────────────────────────
...
  /api/matches/{matchId}/compliments:
    post:
      tags: [칭찬]
      summary: 경기 후 칭찬 Bulk 등록
      description: |
        경기 종료 후 같이 뛴 참가자들을 한 번에 칭찬합니다.

        **칭찬 태그:**
        `MANNERS`(매너 좋아요) | `SKILL`(실력 좋아요) | `PUNCTUAL`(시간 약속 잘 지켜요) |
        `PASSIONATE`(열정적이에요) | `MOOD_MAKER`(분위기 메이커)

        **검증 규칙:**
        - 호출자(rater)가 해당 경기의 APPROVED 참가자여야 함 → 아니면 403
        - 각 ratee도 해당 경기의 APPROVED 참가자여야 함 → 아니면 403
        - 자기 자신 칭찬 불가 → 400
        - 같은 batch 내 rateeId 중복 불가 → 400
        - 이미 해당 경기에서 그 대상을 칭찬했으면 409

        **트랜잭션:** 1명이라도 실패하면 전체 롤백
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ComplimentBulkRequest'
            example:
              compliments:
                - rateeId: 5
                  tags: ["MANNERS", "PUNCTUAL"]
                  comment: "오늘 정말 즐거웠습니다"
                - rateeId: 7
                  tags: ["SKILL"]
                  comment: null
      responses:
        '200':
          description: 칭찬 등록 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "2명에게 칭찬을 보냈습니다"
                data:
                  created: 2
        '400':
          description: 자기 자신 칭찬 또는 중복 rateeId
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                selfCompliment:
                  value: { success: false, message: "자기 자신은 칭찬할 수 없습니다.", data: null }
                duplicate:
                  value: { success: false, message: "이미 칭찬한 참가자가 포함되어 있습니다.", data: null }
        '403':
          description: 경기 참가자가 아님 (rater 또는 ratee)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example: { success: false, message: "경기에 참가하지 않은 회원입니다.", data: null }
  /api/members/{memberId}/compliments:
    get:
      tags: [칭찬]
      summary: 특정 회원이 받은 칭찬 목록 (Cursor 페이지네이션)
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: cursor
          in: query
          schema:
            type: integer
            format: int64
        - name: size
          in: query
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "칭찬 목록 조회 성공"
                data:
                  items:
                    - complimentId: 88
                      matchId: 12
                      raterId: 3
                      rateeId: 1
                      tags: ["MANNERS", "PUNCTUAL"]
                      comment: "오늘 정말 즐거웠습니다"
                      createdAt: "2026-05-17T19:30:00"
                    - complimentId: 87
                      matchId: 10
                      raterId: 4
                      rateeId: 1
                      tags: ["SKILL"]
                      comment: null
                      createdAt: "2026-05-15T20:15:00"
                  nextCursor: 87
  /api/members/{memberId}/compliments/summary:
    get:
      tags: [칭찬]
      summary: 회원 칭찬 누적 카운트 (프로필용)
      description: 태그별 칭찬 누적 횟수를 조회합니다. `member_compliment_summary` 테이블에서 단건 조회합니다.
      parameters:
        - name: memberId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "칭찬 통계 조회 성공"
                data:
                  memberId: 1
                  totalCount: 35
                  tagCounts:
                    MANNERS: 12
                    SKILL: 8
                    PUNCTUAL: 7
                    PASSIONATE: 5
                    MOOD_MAKER: 3
  /api/matches/{matchId}/compliments/me:
    get:
      tags: [칭찬]
      summary: 특정 경기에서 내가 한/받은 칭찬 조회
      parameters:
        - name: matchId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
              example:
                success: true
                message: "경기 칭찬 조회 성공"
                data:
                  given:
                    - complimentId: 88
                      matchId: 12
                      raterId: 1
                      rateeId: 5
                      tags: ["MANNERS"]
                      comment: "재밌었어요"
                      createdAt: "2026-05-17T19:30:00"
                  received:
                    - complimentId: 90
                      matchId: 12
                      raterId: 4
                      rateeId: 1
                      tags: ["SKILL", "PASSIONATE"]
                      comment: null
                      createdAt: "2026-05-17T19:35:00"

# ============================================================
# Components
# ============================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "카카오 로그인 후 발급받은 accessToken"
  responses:
    Unauthorized:
      description: 인증 실패 (토큰 없음 또는 만료)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            message: "인증이 필요합니다."
            data: null
...
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          nullable: true
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
        data:
          nullable: true
          example: null
...
    ComplimentSubmitItem:
      type: object
      required: [rateeId, tags]
      properties:
        rateeId:
          type: integer
          format: int64
        tags:
          type: array
          minItems: 1
          maxItems: 5
          uniqueItems: true
          items:
            type: string
            enum: [MANNERS, SKILL, PUNCTUAL, PASSIONATE, MOOD_MAKER]
        comment:
          type: string
          maxLength: 200
          nullable: true
    ComplimentBulkRequest:
      type: object
      required: [compliments]
      properties:
        compliments:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/ComplimentSubmitItem'

    # ── 응답 스키마 ──────────────────────────────────
...